// typechain/contracts/GoldenBounty.sol/GoldenBounty.ts
export interface QuestionCreatedEventObject {
  questionAddress: string;
  subjectUUID: string;
  predicateUUID: string;
}
export type QuestionCreatedEvent = TypedEvent<
  [string, string, string],
  QuestionCreatedEventObject
>;